revert spectest exports.
